Update TrailerRCJ model and AWS Lambda layer details
The commit updates the TrailerRCJ model in the schema.graphql file,
adjusting the @auth rules and fixing a syntax error. It also modifies
the AWS Lambda layer version details in the
awscloudformation-template.json file, updating the resource names and
references. Additionally, the description of the layer version in the
parameters.json file is updated.

Other changes include:
- Simplifying code in ViewButtonAWS.tsx by using optional chaining and
destructuring assignments.
- Removal of the EditIcon import in ChassisTable.tsx.
- Refactoring of the useAuth function in the utils.js file to handle
authentication events and fetch user attributes.